requirements.txt
# YouTube API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Timezone support
pytz>=2023.3

# Utilities
python-dotenv>=1.0.0
